requirements.txt
telethon>=1.34
python-dotenv>=1.0
aiofiles>=23.2
faiss-cpu>=1.7
sentence-transformers>=2.6
numpy>=1.26
python-dateutil>=2.8